renamed component and removed unused function
